Replace participant filter panel
