+++ src/main/java/de/vinado/wicket/participate/ui/event/details/DetailedParticipantFilterPanel.java
@@ -2,16 +2,14 @@ package de.vinado.wicket.participate.ui.event.details;
 
 import de.vinado.wicket.participate.model.Event;
 import de.vinado.wicket.participate.model.filters.ParticipantFilter;
-import org.apache.wicket.Component;
-import org.apache.wicket.event.Broadcast;
+import de.vinado.wicket.participate.ui.event.ParticipantFilterPanel;
 import org.apache.wicket.markup.html.WebMarkupContainer;
-import org.apache.wicket.markup.html.panel.GenericPanel;
 import org.apache.wicket.model.IModel;
 
 /**
  * @author Vincent Nadoll (vincent.nadoll@gmail.com)
  */
-public abstract class DetailedParticipantFilterPanel extends GenericPanel<ParticipantFilter> {
+public abstract class DetailedParticipantFilterPanel extends ParticipantFilterPanel {
 
     private final IModel<Event> eventModel;
 
@@ -21,13 +19,7 @@ public abstract class DetailedParticipantFilterPanel extends GenericPanel<Partic
     }
 
     @Override
-    protected void onInitialize() {
-        super.onInitialize();
-
-        add(form("form"));
-    }
-
-    private WebMarkupContainer form(String id) {
+    protected WebMarkupContainer form(String id) {
         return new DetailedParticipantFilterForm(id, getModel(), eventModel) {
             @Override
             protected void onApply() {
@@ -35,12 +27,6 @@ public abstract class DetailedParticipantFilterPanel extends GenericPanel<Partic
             }
         };
     }
-
-    private void onApply() {
-        send(getScope(), Broadcast.BREADTH, new ParticipantFilterIntent(getModelObject()));
-    }
-
-    protected abstract Component getScope();
 
     @Override
     protected void onDetach() {
