+++ src/main/resources/de/vinado/wicket/participate/ui/event/EventPanel.html
@@ -35,6 +35,7 @@
                     <dd wicket:id="description"></dd>
                 </wicket:enclosure>
             </dl>
+            <div class="clearfix" style="margin-top: 5px; margin-bottom: 5px; border-bottom: 1px solid #e5e5e5;"></div>
                 <div wicket:id="filterPanel"></div>
                 <div class="table-responsive">
                     <table wicket:id="dataTable"></table>
