+++ src/main/resources/de/vinado/wicket/participate/ui/event/ParticipantFilterPanel.html
@@ -5,11 +5,7 @@
 </head>
 <body>
 <wicket:panel>
-    <wicket:extend>
-        <input wicket:id="name">
-        <select wicket:id="invitationStatus"></select>
-        <select wicket:id="voice"></select>
-    </wicket:extend>
+    <form wicket:id="form"/>
 </wicket:panel>
 </body>
 </html>
