--- src/main/resources/de/vinado/wicket/participate/ui/event/details/DetailedParticipantFilterPanel.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:wicket="http://wicket.apache.org">
-<head>
-    <meta charset="utf-8"/>
-</head>
-<body>
-<wicket:panel>
-    <form wicket:id="form"/>
-</wicket:panel>
-</body>
-</html>
